+++ testing/twk-multi.js
@@ -257,6 +257,18 @@ TWKMulti.prototype.getInfo = function(info) {
 		info.doMulti = false;
 	}
 };
+TWKMulti.prototype.isCustomGamma = function() {
+	return false;
+};
+TWKMulti.prototype.isCustomGamut = function() {
+	var tweaks = this.inputs.tweaks.checked;
+	var tweak = this.tweakCheck.checked;
+	if (tweaks && tweak) {
+		return true;
+	} else {
+		return false;
+	}
+};
 TWKMulti.prototype.events = function() {
 	this.tweakCheck.onclick = function(here){ return function(){
 		here.toggleTweak();