+++ testing/twk-white.js
@@ -336,6 +336,18 @@ TWKWHITE.prototype.getInfo = function(info) {
 		info.doWB = false;
 	}
 };
+TWKWHITE.prototype.isCustomGamma = function() {
+	return false;
+};
+TWKWHITE.prototype.isCustomGamut = function() {
+	var tweaks = this.inputs.tweaks.checked;
+	var tweak = this.tweakCheck.checked;
+	if (tweaks && tweak) {
+		return true;
+	} else {
+		return false;
+	}
+};
 TWKWHITE.prototype.events = function() {
 	this.tweakCheck.onclick = function(here){ return function(){
 		here.toggleTweak();